pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sentimentscribe"
version = "1.0.0"
description = "Production-ready book review sentiment analysis system"
authors = [{ name = "Shah-king" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["sentiment-analysis", "nlp", "machine-learning", "mlops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scipy>=1.12.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "pydantic>=2.6.0",
    "mlflow>=2.11.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[project.optional-dependencies]
transformer = [
    "torch>=2.2.0",
    "transformers>=4.39.0",
    "datasets>=2.18.0",
]
monitoring = [
    "evidently>=0.4.20",
]
explainability = [
    "shap>=0.45.0",
]
dashboard = [
    "streamlit>=1.32.0",
    "plotly>=5.20.0",
    "requests>=2.31.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]
all = [
    "sentimentscribe[transformer,monitoring,explainability,dashboard,dev]",
]

[project.scripts]
senti-train = "train:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short -v"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
